feat(nvim): replace "packer" in favour of "lazy" as plugin manager
reduces the start-up time by up to 50%

### Benchmark

#### Packer

Time (mean ± σ):      1.201 s ±  0.252 s    [User: 0.373 s, System: 0.528 s]
Range (min … max):    0.895 s …  1.537 s    10 runs

#### Lazy

Time (mean ± σ):     614.7 ms ± 110.1 ms    [User: 274.5 ms, System: 250.5 ms]
Range (min … max):   501.9 ms … 824.4 ms    10 runs
