+++ .gitignore
@@ -352,7 +352,6 @@ zsdoc/data
 !/tests/_output/.gitkeep
 
 # End of https://www.toptal.com/developers/gitignore/api/visualstudiocode,vim,ruby,rails,sass,macos,node,ssh,zsh
-packer_compiled.lua
 nvim/nvim-macos/
 
 # asdf runtime manager
