--- nvim/after/plugin/text-case.lua
@@ -1,8 +1,0 @@
-local found_textcase, textcase = pcall(require, "textcase")
-if not found_textcase then
-  return
-end
-
-textcase.setup({})
-
-vim.keymap.set({ "n", "v" }, "<leader>cc", "<cmd>TextCaseOpenTelescope<CR>", { desc = "Telescope" })
