+++ nvim/lua/cange/init.lua
@@ -1,9 +1,6 @@
--- order is important
-Cange = require("cange.utils")
--- secondary
+-- [[ Other stuff ]]
 Cange.reload("cange.cmp")
 Cange.reload("cange.core")
 Cange.reload("cange.keymaps")
 Cange.reload("cange.lsp")
-Cange.reload("cange.plugins")
 Cange.reload("cange.telescope")
