+++ nvim/lua/plugins/fidget.lua
@@ -1,0 +1,10 @@
+return {
+  "j-hui/fidget.nvim", -- shows LSP initialization progress
+  config = function()
+    require("fidget").setup({
+      text = {
+        spinner = "dots",
+      },
+    })
+  end,
+}
