+++ nvim/lua/plugins/text-case.lua
@@ -1,0 +1,8 @@
+return {
+  "johmsalas/text-case.nvim", -- text case converter (camel case, etc.,
+  config = function()
+    require("textcase").setup()
+
+    vim.keymap.set({ "n", "v" }, "<leader>cc", "<cmd>TextCaseOpenTelescope<CR>", { desc = "Telescope" })
+  end,
+}
