+++ nvim/lua/plugins/web-devicons.lua
@@ -1,0 +1,13 @@
+return {
+  "kyazdani42/nvim-web-devicons", -- File icons
+  config = function()
+    require("nvim-web-devicons").setup({
+      -- your personnal icons can go here (to override)
+      -- DevIcon will be appended to `name`
+      override = {},
+      -- globally enable default icons (default to false)
+      -- will get overriden by `get_icons` option
+      default = true,
+    })
+  end,
+}
